tilegrid: separate tilegrid and address generation
